moved db related files into separate folder
